# contracts/openapi/torgnexa-v1.yaml
openapi: 3.1.0
info:
  title: TORGNEXA API
  version: 0.16.0
servers:
  - url: /api/v1
paths:
...
  /settings/security/configuration:
    get:
      operationId: getSettingsSecurityConfiguration
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Runtime OIDC configuration state without claiming provider health, content: {application/json: {schema: {$ref: '#/components/schemas/SettingsSecurityConfiguration'}}}}
        '403': {description: Caller lacks settings.security.read}
  /settings/security/sessions:
    get:
      operationId: listSettingsSecuritySessions
      security: [{bearerAuth: []}]
      parameters:
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: cursor, in: query, required: false, schema: {type: string, maxLength: 64}}
      responses:
        '200': {description: Minimized tenant-scoped OIDC sessions observed by TORGNEXA, content: {application/json: {schema: {$ref: '#/components/schemas/SettingsSecuritySessionPage'}}}}
        '403': {description: Caller lacks settings.security.read}
...
  /settings/security/logins:
    get:
      operationId: listSettingsSecurityLogins
      security: [{bearerAuth: []}]
      parameters:
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: cursor, in: query, required: false, schema: {type: string, maxLength: 64}}
      responses:
        '200': {description: Cursor-paginated TORGNEXA-observed OIDC session history, content: {application/json: {schema: {$ref: '#/components/schemas/SettingsSecurityLoginPage'}}}}
        '403': {description: Caller lacks settings.security.read}
  /settings/security/audit:
    get:
      operationId: listSettingsSecurityAudit
      security: [{bearerAuth: []}]
      parameters:
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: cursor, in: query, required: false, schema: {type: string, maxLength: 64}}
      responses:
        '200': {description: Cursor-paginated settings mutation evidence with actor and correlation, content: {application/json: {schema: {$ref: '#/components/schemas/AuditPage'}}}}
        '403': {description: Caller lacks settings.security.read}
...
  /inventory/positions/{position_id}:
    parameters: [{name: position_id, in: path, required: true, schema: {type: string}}]
    get:
      operationId: getInventoryPosition
      responses: {'200': {description: Position card and immutable movement history}, '404': {description: Position not found}}
...
  /inventory/positions:
    get:
      operationId: listInventoryPositions
      responses:
        '200':
          description: Tenant-scoped canonical inventory positions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 100
                    items:
                      type: object
                      required: [ID, SKU, ProductTitle, WarehouseCode, WarehouseName, OnHand, Reserved, Available, Unit, UpdatedAt]
                      properties:
                        ID: {type: string}
                        SKU: {type: string}
                        ProductTitle: {type: string}
                        WarehouseCode: {type: string}
                        WarehouseName: {type: string}
                        OnHand: {type: string}
                        Reserved: {type: string}
                        Available: {type: string}
                        Unit: {type: string}
                        UpdatedAt: {type: string, format: date-time}
        '401': {description: Authenticated tenant scope missing}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    SettingsSecurityConfiguration:
      type: object
      additionalProperties: false
      required: [provider, issuer, client_id, configuration_status, configuration_source, provider_health]
      properties:
        provider: {type: string, const: oidc}
        issuer: {type: string, format: uri}
        client_id: {type: string, minLength: 1, maxLength: 256}
        configuration_status: {type: string, enum: [configured]}
        configuration_source: {type: string, enum: [runtime]}
        provider_health: {type: string, enum: [not_verified]}
    SettingsSecuritySession:
      type: object
      additionalProperties: false
      required: [session_ref, user_ref, status, client_kind, authenticated_at, first_seen_at, last_seen_at, expires_at, current]
      properties:
        session_ref: {type: string, pattern: '^[0-9a-f]{64}$'}
        user_ref: {type: string, pattern: '^[0-9a-f]{12}$'}
        status: {type: string, enum: [active, revoked]}
        client_kind: {type: string, enum: [browser, mobile, api, unknown]}
        authenticated_at: {type: string, format: date-time}
        first_seen_at: {type: string, format: date-time}
        last_seen_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        revoked_at: {type: string, format: date-time}
        current: {type: boolean}
    SettingsSecuritySessionPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/SettingsSecuritySession'}}
        next_cursor: {type: string, maxLength: 64}
...
    SettingsSecurityLogin:
      type: object
      additionalProperties: false
      required: [id, session_ref, event_type, client_kind, occurred_at]
      properties:
        id: {type: string, format: uuid}
        session_ref: {type: string, pattern: '^[0-9a-f]{64}$'}
        event_type: {type: string, enum: [session_observed, session_revoked]}
        client_kind: {type: string, enum: [browser, mobile, api, unknown]}
        occurred_at: {type: string, format: date-time}
    SettingsSecurityLoginPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, evidence_scope]
      properties:
        items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/SettingsSecurityLogin'}}
        next_cursor: {type: string, maxLength: 64}
        evidence_scope: {type: string, const: torgnexa_observed_oidc_sessions}
    AuditRecord:
      type: object
      additionalProperties: false
      required: [id, actor_id, source, action, resource_type, resource_id, correlation_id, risk, summary, created_at]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        actor_id: {type: string, minLength: 1, maxLength: 255}
        source: {type: string, minLength: 1, maxLength: 64}
        action: {type: string, minLength: 1, maxLength: 160}
        resource_type: {type: string, minLength: 1, maxLength: 128}
        resource_id: {type: string, minLength: 1, maxLength: 512}
        correlation_id: {type: string, minLength: 1, maxLength: 128}
        risk: {type: string, enum: [read, write_safe, write_sensitive, legally_significant]}
        summary: {type: object, additionalProperties: true}
        created_at: {type: string, format: date-time}
    AuditPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/AuditRecord'}}
        next_cursor: {type: string, maxLength: 64}
...